+++ src/File/Collection/DirectoryCollection.php
@@ -2,12 +2,12 @@
 
 namespace LDL\File\Collection;
 
-use LDL\File\Helper\FileTypeHelper;
+use LDL\File\Constants\FileTypeConstants;
 use LDL\File\Validator\FileTypeValidator;
-use LDL\Type\Collection\AbstractCollection;
+use LDL\Type\Collection\AbstractTypedCollection;
 use LDL\Type\Collection\Traits\Validator\AppendValueValidatorChainTrait;
 
-final class DirectoryCollection extends AbstractCollection
+final class DirectoryCollection extends AbstractTypedCollection
 {
     use AppendValueValidatorChainTrait;
 
@@ -17,7 +17,7 @@ final class DirectoryCollection extends AbstractCollection
 
         $this->getAppendValueValidatorChain()
             ->getChainItems()
-            ->append(new FileTypeValidator([FileTypeHelper::FILE_TYPE_DIRECTORY]))
+            ->append(new FileTypeValidator([FileTypeConstants::FILE_TYPE_DIRECTORY]))
             ->lock();
     }
 }