+++ src/File/Collection/JsonFileCollection.php
@@ -5,10 +5,10 @@ namespace LDL\File\Collection;
 use LDL\File\Validator\FileExistsValidator;
 use LDL\File\Validator\JsonFileValidator;
 use LDL\File\Validator\ReadableFileValidator;
-use LDL\Type\Collection\AbstractCollection;
+use LDL\Type\Collection\AbstractTypedCollection;
 use LDL\Type\Collection\Traits\Validator\AppendValueValidatorChainTrait;
 
-final class JsonFileCollection extends AbstractCollection
+final class JsonFileCollection extends AbstractTypedCollection
 {
     use AppendValueValidatorChainTrait;