+++ src/File/Collection/ReadableFileCollection.php
@@ -4,10 +4,10 @@ namespace LDL\File\Collection;
 
 use LDL\File\Validator\FileExistsValidator;
 use LDL\File\Validator\ReadableFileValidator;
-use LDL\Type\Collection\AbstractCollection;
+use LDL\Type\Collection\AbstractTypedCollection;
 use LDL\Type\Collection\Traits\Validator\AppendValueValidatorChainTrait;
 
-final class ReadableFileCollection extends AbstractCollection
+final class ReadableFileCollection extends AbstractTypedCollection
 {
     use AppendValueValidatorChainTrait;