+++ src/File/Collection/WritableFileCollection.php
@@ -4,10 +4,10 @@ namespace LDL\File\Collection;
 
 use LDL\File\Validator\FileExistsValidator;
 use LDL\File\Validator\WritableFileValidator;
-use LDL\Type\Collection\AbstractCollection;
+use LDL\Type\Collection\AbstractTypedCollection;
 use LDL\Type\Collection\Traits\Validator\AppendValueValidatorChainTrait;
 
-final class WritableFileCollection extends AbstractCollection
+final class WritableFileCollection extends AbstractTypedCollection
 {
     use AppendValueValidatorChainTrait;